requirements.txt
# core
pandas>=2.2
requests>=2.32
python-dateutil>=2.9
tqdm>=4.66
pydantic>=2.7
streamlit>=1.36
python-dotenv

# gridstatus + numpy pins for Py 3.12
gridstatusio==0.14.0
numpy==1.26.4
